Well, it looks crappy, but it should now be impossible to get stuck with the blocks on level 3.
They can no longer be pushed into the left corner, and even if they barely push the blocks over the edge, they will far just far enough for the player to slip behind them.

I was thinking redoing the art for the blocks so they are sloped, but then it wouldn't make sense for the player to be able to get behind them.