Whole bunch of stuff for the boss fight.
Added several new scripts that manage the shots of both the player and the boss, new boss shot sprites+animations.

I remade the bird since something was wrong with the previous version of it.
Added movement and killing on touch to the boss.
Movement is based on Perlin noise, with variables that can be adjusted in the inspector.
Zoomed out the camera a bit.